Moved SocketType enum to config
